requirements.txt
fabric
gitpython